bw/pet: filter out negative health values

+++ BinWeevils.Protocol/Form/Pet/UpdatePetStats.cs
@@ -8,7 +8,7 @@ namespace BinWeevils.Protocol.Form.Pet
         [PropertyShape(Name = "petID")] public uint m_petID;
         [PropertyShape(Name = "mentalEnergy")] public byte m_mentalEnergy;
         [PropertyShape(Name = "fuel")] public byte m_fuel;
-        [PropertyShape(Name = "health")] public byte m_health;
+        [PropertyShape(Name = "health")] public short m_health;
         [PropertyShape(Name = "fitness")] public byte m_fitness;
         [PropertyShape(Name = "experience")] public uint m_experience;
     }
